GT-1 - Extracted ErrorCode
